made ops file

+++ frog/gradcheck.py
@@ -7,7 +7,7 @@ def jacobian(model, input):
 
   ji = input.data.reshape(-1).shape[-1]  # jacobian of input
   jo = output.data.reshape(-1).shape[-1] # jacobian of output
-  J = np.zeros((jo, ji))
+  J = np.zeros((jo, ji), dtype=np.float32)
 
   for o in range(jo):
     o_scalar = Tensor(mask_like(output.data, o, 1.)).mul(output).sum()
@@ -33,7 +33,7 @@ def numerical_jacobian(model, input, eps = 1e-6):
 
   ji = input.data.reshape(-1).shape[-1]
   jo = output.data.reshape(-1).shape[-1]
-  NJ = np.zeros((jo, ji))
+  NJ = np.zeros((jo, ji), dtype=np.float32)
 
   for o in range(jo):
     for i in range(ji):
